+++ src/components/settings/SettingsItemInput/index.ts
@@ -1,0 +1,1 @@
+export { SettingsItemInput } from "./SettingsItemInput.tsx"